requirements.txt
python-telegram-bot>=20.6
yt-dlp
requests
Pillow
opencv-python
moviepy
pyzbar
validators
pytesseract
torch
torchvision
torchaudio
ultralytics
transformers
diffusers
accelerate
scipy
numpy
psutil
diskcache
fsspec
networkx
jinja2
sympy
mpmath
aiohttp
python-dotenv
